refactor(interactivity): extract resolveHover/tooltips from domain-utils
Move the hover + standard/manual tooltip binders into
src/interactivity/tooltips.ts and drop the now-unused diagnostics +
tooltip-type imports from domain-utils. No behaviour change.
