+++ src/domain-utils.ts
@@ -2,8 +2,6 @@
 import powerbi from 'powerbi-visuals-api';
 import IVisualHost = powerbi.extensibility.visual.IVisualHost;
 import ISelectionId = powerbi.visuals.ISelectionId;
-import TooltipShowOptions = powerbi.extensibility.TooltipShowOptions;
-import VisualTooltipDataItem = powerbi.extensibility.VisualTooltipDataItem;
 
 // External dependencies
 import { select, Selection } from 'd3-selection';
@@ -28,8 +26,6 @@ import {
 } from './sanitize';
 import { CONTENT_TOKEN, ROW_TOKEN, substitute } from './template-engine';
 import { buildHighlightedFragment } from './diagnostics/highlight-html';
-import { recordTooltipEvent } from './diagnostics/event-recorder';
-import { tooltipContext, TooltipItem } from './diagnostics/host-events';
 
 // Re-export sanitize pipeline entry points so existing callers that import
 // from './domain-utils' continue to work after the Task 7 extraction.
@@ -526,133 +522,6 @@ export function resolveScrollableContent(
         scrollbars: {
             clickScrolling: true
         }
-    });
-}
-
-/**
- * Handle eventing when a data element is hovred over. This includes showing
- * the tooltip and toggling appropriate class names for style hooks.
- *
- * @param dataElements      - The elements to analyse and process.
- * @param host              - Visual host services.
- * @param hasGranularity    - Whether we have granularity or not.
- */
-export function resolveHover(
-    dataElements: Selection<any, IHtmlEntry, any, any>,
-    host: IVisualHost,
-    hasGranularity: boolean
-) {
-    bindStandardTooltips(dataElements, host, hasGranularity);
-    bindManualTooltips(dataElements, host);
-}
-
-/**
- * If we don't have any granularity, we will look for elements that have
- * a tooltip attribute and use this to show the tooltip.
- *
- * @param dataElements      - The elements to analyse and process.
- * @param host              - Visual host services.
- */
-function bindManualTooltips(
-    dataElements: Selection<any, IHtmlEntry, any, any>,
-    host: IVisualHost
-) {
-    const { tooltipService } = host;
-    const {
-        manualTooltipSelector,
-        manualTooltipDataPrefix,
-        manualTooltipDataTitle,
-        manualTooltipDataValue
-    } = VisualConstants.dom;
-    const manualTooltipElements = dataElements.selectAll(
-        `.${manualTooltipSelector}`
-    );
-    const titleExp = new RegExp(
-        `${manualTooltipDataPrefix}${manualTooltipDataTitle}`,
-        'g'
-    );
-    const valueExp = new RegExp(
-        `${manualTooltipDataPrefix}${manualTooltipDataValue}`,
-        'g'
-    );
-    manualTooltipElements.on('mouseover mousemove', (event) => {
-        const dataset = event.currentTarget.dataset;
-        const keys = Object.keys(dataset).map((key) =>
-            key.replace(titleExp, '').replace(valueExp, '')
-        );
-        const uniqueKeys = [...new Set(keys)];
-        const dataItems: VisualTooltipDataItem[] = uniqueKeys.map((key) => ({
-            displayName:
-                dataset[
-                    `${manualTooltipDataPrefix}${manualTooltipDataTitle}${key}`
-                ] || '',
-            value:
-                dataset[
-                    `${manualTooltipDataPrefix}${manualTooltipDataValue}${key}`
-                ] || ''
-        }));
-        if (dataItems.length > 0) {
-            const options: TooltipShowOptions = {
-                coordinates: [event.clientX, event.clientY],
-                isTouchEvent: true,
-                dataItems,
-                identities: []
-            };
-            tooltipService.show(options);
-            recordTooltipEvent(
-                'show',
-                'manual',
-                tooltipContext(dataItems as TooltipItem[])
-            );
-        }
-    });
-    manualTooltipElements.on('mouseout', () => {
-        tooltipService.hide({ immediately: true, isTouchEvent: true });
-        recordTooltipEvent('hide', 'manual', '');
-    });
-}
-
-/**
- * For standard data elements, working with the data roles and correct
- * rules, we will apply the regular tooltip handling.
- *
- * @param dataElements      - The elements to analyse and process.
- * @param host              - Visual host services.
- * @param hasGranularity    - Whether we have granularity or not.
- */
-function bindStandardTooltips(
-    dataElements: Selection<any, IHtmlEntry, any, any>,
-    host: IVisualHost,
-    hasGranularity: boolean
-) {
-    const { tooltipService } = host;
-    dataElements.on('mouseover mousemove', (event, d) => {
-        select(event.currentTarget).classed(
-            VisualConstants.dom.hoverClassSelector,
-            true
-        );
-        if (hasGranularity || d.tooltips.length > 0) {
-            const options: TooltipShowOptions = {
-                coordinates: [event.clientX, event.clientY],
-                isTouchEvent: true,
-                dataItems: d.tooltips,
-                identities: [d.identity]
-            };
-            tooltipService.show(options);
-            recordTooltipEvent(
-                'show',
-                'contextual',
-                tooltipContext(d.tooltips as TooltipItem[])
-            );
-        }
-    });
-    dataElements.on('mouseout', (event) => {
-        select(event.currentTarget).classed(
-            VisualConstants.dom.hoverClassSelector,
-            false
-        );
-        tooltipService.hide({ immediately: true, isTouchEvent: true });
-        recordTooltipEvent('hide', 'contextual', '');
     });
 }
