+++ src/interactivity/index.ts
@@ -1,3 +1,4 @@
 // Public interactivity surface. Consumers import from './interactivity'
 // rather than reaching into individual modules.
 export { BehaviorManager, IHtmlBehaviorOptions } from './behavior';
+export { resolveHover } from './tooltips';
